refactor(): Rename App component to Layout
