+++ resources/js/components/Layout.vue
